Remove definition of _diffrn_radiation.diffrn_id
This is defined in mmCIF.dic, so is available if someone really
wants it. However, it does not fit into our concept of associating
a single radiation source with a particular experiment; instead
it associates a particular experiment with a particular radiation
source, which can only make sense when there is a one-to-one
relationship between source and experiment. As many measurements
can be performed on a single source, this dataname will rarely
be useful.